bench/pyproject.toml
[project]
name = "tq-vlm-bench"
version = "0.1.0"
description = "Benchmark framework for TurboQuant KV cache experiments on llama.cpp VLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.0.0",
    "gguf>=0.18.0",
    "httpx>=0.28.0",
    "jupyterlab>=4.0.0",
    "matplotlib>=3.9.0",
    "pandas>=2.2.0",
    "pillow>=11.0.0",
    "pytest>=9.0.3",
    "pyyaml>=6.0.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.0",
    "tqdm>=4.66.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tq_bench"]

[tool.uv]
package = true

